- Progress bar and message on startup screen is no longer static, now displays real messages and progress the log (dynamically) - Moved some functions from start.java to misc.java (get date of compiled class) - Minor changes to make the log more pretty - The intention of adding a dynamic startup screen is to enable the reading of non-cached SPDX documents when starting up the program. This way the end-user won't get worried if suddenly takes a bit more of time to finish loading. I'm kind of proud for this feature. First time getting this kind of startup bars working so neatly. :-)

+++ src/script/log.java
@@ -115,7 +115,7 @@ public class log {
      * Gets the current number of indexed log messages
      * @return number of messages indexed
      */
-    public static long getCounter() {
+    public static int getCounter() {
         return list.size();
     }
